Extract react datasheet into separate component
